fixing double cast and format
git-svn-id: svn://svn.open-ils.org/ILS/trunk@265 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ OpenSRF/src/libtransport/generic_utils.c
@@ -19,7 +19,7 @@ void get_timestamp( char buf_36chars[]) {
 double get_timestamp_millis() {
 	struct timeb t;
 	ftime(&t);
-	double time	= ( (int)t.time	+ ( ((float)t.millitm) / 1000 ) );
+	double time	= ( (int)t.time	+ ( ((double)t.millitm) / 1000 ) );
 	return time;
 }
